# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ThankATech API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /technicians:
    get:
      operationId: listTechnicians
      tags: [technicians]
      summary: List/search technicians
      parameters:
        - in: query
          name: specialty
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of technicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Technician"
    post:
      operationId: createTechnician
      tags: [technicians]
      summary: Create technician profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TechnicianInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
  /technicians/{id}:
    get:
      operationId: getTechnician
      tags: [technicians]
      summary: Get technician by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Technician
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Technician"
        "404":
          description: Not found
  /technicians/{id}/wall-of-thanks:
    get:
      operationId: getTechnicianWallOfThanks
      tags: [technicians]
      summary: Get technician's wall of thanks
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of thank messages for technician
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThankMessage"
  /technicians/{id}/stats:
    get:
      operationId: getTechnicianStats
      tags: [technicians]
      summary: Get technician stats summary
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TechnicianStats"
...
  /thanks:
    get:
      operationId: listThankMessages
      tags: [thanks]
      summary: List thank messages
      parameters:
        - in: query
          name: technicianId
          schema:
            type: integer
        - in: query
          name: customerId
          schema:
            type: integer
      responses:
        "200":
          description: List of thank messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ThankMessage"
    post:
      operationId: createThankMessage
      tags: [thanks]
      summary: Send a thank you message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ThankMessageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ThankMessage"
...
components:
...
  schemas:
...
    Technician:
      type: object
      required: [id, fullName, specialty, serviceArea, bio, totalThanks, totalEarned, avatarUrl]
      properties:
        id:
          type: integer
        fullName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        specialty:
          type: string
        specialties:
          type: array
          items:
            type: string
        serviceArea:
          type: string
        bio:
          type: string
        hourlyRate:
          type: ["number", "null"]
        certifications:
          type: array
          items:
            type: string
        totalThanks:
          type: integer
        totalEarned:
          type: number
        createdAt:
          type: string
    TechnicianInput:
      type: object
      required: [fullName, specialty, serviceArea, bio]
      properties:
        fullName:
          type: string
        avatarUrl:
          type: string
        specialty:
          type: string
        specialties:
          type: array
          items:
            type: string
        serviceArea:
          type: string
        bio:
          type: string
        hourlyRate:
          type: number
        certifications:
          type: array
          items:
            type: string
    TechnicianStats:
      type: object
      required: [technicianId, totalThanks, totalTips, totalEarned, totalJobs, avgTipAmount]
      properties:
        technicianId:
          type: integer
        totalThanks:
          type: integer
        totalTips:
          type: integer
        totalEarned:
          type: number
        totalJobs:
          type: integer
        avgTipAmount:
          type: number
...
    ThankMessage:
      type: object
      required: [id, jobId, customerId, technicianId, message, tipAmount, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        technicianId:
          type: integer
        technicianName:
          type: string
        technicianAvatar:
          type: ["string", "null"]
        message:
          type: string
        tipAmount:
          type: number
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ThankMessageInput:
      type: object
      required: [jobId, customerId, technicianId, message]
      properties:
        jobId:
          type: integer
        customerId:
          type: integer
        technicianId:
          type: integer
        message:
          type: string
        tipAmount:
          type: number
        photoUrl:
          type: string
...